requirements.txt
PyInstaller
tkinter
pyperclip
requests
flask
python-dotenv
waitress  # Production WSGI server